EXERCISE STILL HAS ERROR IN THE MODEL!
Consider the data set in heart.csv (data description in heart.rtf).
6. Select a suitable prior for the parameters in a logistic regression.
7. Adjust a Bayesian logistic regression model using target has the dependent variable,
using all other variables as predictors.
8. Ascertain the convergence of the Markov chains used in the inference.
9. Generate, using the covariates in the dataset, a “new” dataset using the posterior
distributions for the logistic regression and calculate the mean of these generated
datasets. Repeat this process n times, comparing the histogram of the generated means
to the true population mean.
10. Calculate the Bayes factor for the alternative hypothesis that age increases the
probability of heart failure.